The exact 1-stage function, randomized

This function calculates sample sizes of an exact randomized 1-stage design, for p0<pa

exact1stage(p0, pa, alpha, beta, eps = 0.005, alloc, ...)

Arguments

p0
uninteresting response (null hypothesis H0)

pa
interesting response (alternative hypothesis Ha)

alpha
P(reject H0|H0)

beta
P(reject Ha|Ha)

eps
tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005

alloc
allocation ratio (e.g. 2 for 2:1)

...
refers to type="exact" or "normal" for pdf and cdf difference of two binomial variables

Details

n= total number of patients

n_E= number of patients in experimental arm

n_C= number of patients in control arm

if (E-C)<=r --> futility

if (E-C)> r --> efficacy

Note that all sample sizes can be entered, but only calculations for sample sizes, without any decimals (taking into account allocation ratio)

Variable of interest: diff=E-C

alloc is assumed >=1 (more patients in the experimental arm)

Examples

exact1stage(p0=0.45,pa=0.7,alpha=0.05,beta=0.2,alloc=1,type="normal")

#>    n n_E n_C r    d      alpha      beta   p0  pa alpha_param beta_param
#> 1 90  45  45 7 17.8 0.05488447 0.2042675 0.45 0.7        0.05        0.2