p-value for a 2-stage PhII single arm Sargent design

This function calculates the p-value for a 2-stage PhII single arm Fleming's design A stage-wise ordering of the sample space is used: Let m be the stopping stage and x the total number of responses. A trial outcome (m'; x') is at least as extreme (against H0) as the observed trial outcome (m; x), if one of the 3 following conditions is met: (A) m' = m and x'>=x (B) m' = 1, m = 2 and x'>=a (C) m' = 2, m = 1 and x <=r1 in other words: rejection of H0 in the first stage is more extreme

fleming2stage_pval(n1, n2 = NULL, r1, a, k, p0)

Arguments

n1
total number of patients in stage1

n2
total number of patients in stage2

r1
critical value futility for the first stage

a
critical value efficacy for the first stage

k
overall observed responses (must be larger than r1)

p0
true success probability under H0

Value

p-value

References

Mander AP, Thompson SG. Two-stage designs optimal under the alternative hypothesis for phase II cancer clinical trials. Contemporary Clinical Trials 2010;31:572–578 Qin F et al. Optimal, minimax and admissible two-stage design for phase II oncology clinical trials. BMC Medical Research Methodology 2020;20:126 Nhacolo A, Brannath W. Interval and point estimation in adaptive Phase II trials with binary endpoint. Stat Methods Med Res 2019;28:2635-2648

Examples

fleming2stage_pval(n1=21, n2=24, r1=5, a=10,k=17, p0=0.3)
#> [1] 0.178294