The sargent 2-stage function for two randomized arms

This function calculates sample sizes of the sargent 2-stage design (randomized)

including first stage and second stage cutoff values for futility (r1 and r2) and efficacy (s)

sargent2stage_R(
  p0,
  pa,
  alpha,
  beta,
  eta,
  pi,
  eps = 0.005,
  N_min,
  N_max,
  alloc,
  ...
)

Arguments

p0
uninteresting response (null hypothesis H0)

pa
interesting response (alternative hypothesis Ha)

alpha
P(reject H0|H0)

beta
P(reject Ha|Ha)

eta
P(reject Ha|H0)

pi
P(reject H0|Ha)

eps
tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual pi>=pi-eps); default value = 0.005

N_min
minimum sample size value for grid search

N_max
maximum sample size value for grid search

alloc
allocation ratio (e.g. 2 for 2:1)

...
refers to type="exact" or "normal" for pdf and cdf difference of two binomial variables

Value

dataframe with selected scenarios, with following columns:

r1

if (E_1-C_1)<=r1 -> stop for futility

n1_E

Number of patients for experimental arm in stage 1

n1_C

Number of patients for control arm in stage 1

n1

Total number of patients in stage 1

r2

if (E_2+E_1)-(C_2+C_1)<=r1 -> futility

s

if (E_2+E_1)-(C_2+C_1)>=1 -> efficacy

n2_E

Number of patients for experimental arm in stage 1

n2_C

Number of patients for control arm in stage 1

n2

Total number of patients in stage 2

N

Total number of patients

N_E

Total number of patients in experimental arm

N_C

Total number of patients in control arm

EN.p0

expected sample size, under the null hypothesis

PET.p0

robability of terminating the trial at the end of the first stage, under the null hypothesis

MIN

Selected 'minimax' design

OPT

Selected 'optimal' design

ADMISS

Selected 'admissible' design

alpha

calculated alpha for selected design

beta

calculated beta for selected design

eta

calculated eta for selected design

pi

calculated pi for selected design

lambda

calculated lambda for selected design

delta

calculated delta for selected design

alloc

allocation rate

p0

p0 parameter for design

pa

pa parameter for design

alpha_param

alpha parameter for design

beta_param

beta parameter for design

eta_param

eta parameter for design

pi_param

pi parameter for design

Details

E_1= number of successes experimental arm, stage 1

C_1= number of successes control arm, stage 1

E_2= number of successes experimental arm, stage 2

C_2= number of successes control arm, stage 2

with E_i~Bin(n_i,P) and C_i~Bin(n_i,P), i=1,2

if (E_1-C_1)<=r1 -> stop futility

if (E_2+E_1) - (C_2+C_1)<=r2 -> futility

if (E_2+E_1) - (C_2+C_1)>=s -> efficacy

Note that all sample sizes can be entered, but only calculations for sample sizes, without any decimals (taking into account allocation ratio)

Variable of interest: Z=E-C, so always in this direction

alloc >=1 (more patients in the experimental arm)

Examples

# \donttest{
sargent2stage_R(
  p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.15, eta = 0.8, pi = 0.8, eps = 0.005,
  N_min = 88, N_max = 90, alloc = 1, type = "normal"
)

#> Warning: no DISPLAY variable so Tk is not available

#> Error in structure(.External(.C_dotTclObjv, objv), class = "tclObj"): [tcl] invalid command name "toplevel".

# }