Onestage combination function for optimal designs

This function calculates sample sizes of the Fleming 1-stage , Simon-2stage, Sargent's 1- and 2-stage design. calculations are performed jointly for different sets of proportions (p0, pA) and operating characteristics. All designs are summarized in 2 tables, 1 for one-stage designs (Fleming and sargent) and 1 for two-stage designs (Simon and sargent)

allsinglearm(p0, pa, alpha, beta, alpha2 = alpha, beta2, pi, eta, eps = 0.005)

Arguments

p0
uninteresting response (null hypothesis H0), can be a vector

pa
interesting response (alternative hypothesis Ha), can be a vector, always same length as p0. The corresponding elements of p0 and pa are taken as a set of proportions

alpha
P(reject H0|H0) for Fleming and Simon designs, can be a vector

beta
P(reject Ha|Ha) for Fleming and Simon designs, can be a vector

alpha2
P(reject H0|H0) for Sargent designs, can be a vector, same lenth as alpha

beta2
P(reject Ha|Ha) for Sargent designs, can be a vector, same length as beta

pi
P(reject H0|Ha) for Sargent designs, can be a vector, same length as beta2

eta
P(reject Ha|H0) for Sargent designs, can be a vector, same length as alpha2

eps
tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005

Examples

allsinglearm(p0 = 0.1, pa = 0.7,
             alpha = 0.05, beta = 0.2, beta2 = 0.1, pi = 0.8, eta = 0.8)
#> [[1]]
#>      p0  pa alpha beta N_Flem R_Flem alpha beta  pi eta N_Sar S_Sar R_Sar
#> one 0.1 0.7  0.05  0.2      1      4  0.05  0.1 0.8 0.8     3     1     1
#> 
#> [[2]]
#>      p0  pa alpha beta R1_Sim N1_Sim R_sim N2_Sim N_Sim alpha beta  pi eta
#> two 0.1 0.7  0.05  0.2      1      0     2      1     2  0.05  0.1 0.8 0.8
#>     R1_Sar N1_Sar R_Sar S_Sar N2_Sar N_Sar
#> two      1      0     2     1      3     3
#>